board.ts decomposition
